Move slot components to subdirectory, rename aria-labels to block-aria-labels
