+++ src/components/block/block-aria-labels.ts
