+++ src/components/block/slot/ArgumentSlot.tsx
@@ -1,8 +1,8 @@
-import { BlockLabels } from "./aria-labels.ts";
+import { BlockLabels } from "../block-aria-labels.ts";
 import { Box } from "@mui/material";
-import { Block, type BlockProps } from "./Block.tsx";
+import { Block, type BlockProps } from "../Block.tsx";
 import { useDroppable } from "@dnd-kit/core";
-import { Draggable } from "../../common/dnd-wrappers/Draggable.tsx";
+import { Draggable } from "../../../common/dnd-wrappers/Draggable.tsx";
 import type { ReactElement } from "react";
 import { ArgumentSlotStyles } from "./argument-slot-styles.ts";
