+++ src/components/block/slot/PresentationalArgumentSlot.tsx
@@ -1,6 +1,6 @@
-import { BlockLabels } from "./aria-labels.ts";
+import { BlockLabels } from "../block-aria-labels.ts";
 import { Box } from "@mui/material";
-import { Block, type BlockProps } from "./Block.tsx";
+import { Block, type BlockProps } from "../Block.tsx";
 import type { ReactElement } from "react";
 import type { ArgumentSlotProps } from "./ArgumentSlot.tsx";
 import { ArgumentSlotStyles } from "./argument-slot-styles.ts";
